requirements/base.txt
# Core Django and dependencies
Django==5.2.7
python-decouple==3.8

# Database
psycopg2-binary==2.9.10

# Redis for caching and sessions
django-redis==5.4.0
redis==5.2.1

# REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0

# Celery for async tasks
celery==5.4.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.1
Pillow==11.0.0
